flask_azure_app/requirements.txt
Flask
keras
tensorflow
pillow
pandas
matplotlib
sklearn
seaborn
numpy
requests
notebook
ipywidgets
pyarrow
azureml.core
opencv-contrib-python-headless
python-dotenv